pyproject.toml
[project]
name = "instashare-backend-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp==5.2.0",
    "asgiref==3.8.1",
    "astroid==3.3.5",
    "billiard==4.2.0",
    "black",
    "brotli==1.1.0",
    "celery==5.4.0",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cron-descriptor==1.4.5",
    "dill==0.3.9",
    "dj-database-url==2.2.0",
    "django==5.1.1",
    "django-celery-beat==2.7.0",
    "django-cors-headers==4.4.0",
    "django-timezone-field==7.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-yasg==1.21.7",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "inflection==0.5.1",
    "iniconfig==2.0.0",
    "isort==5.13.2",
    "kombu==5.4.1",
    "mccabe==0.7.0",
    "mypy-extensions==1.0.0",
    "packaging==24.1",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "prompt-toolkit==3.0.47",
    "psycopg>=3.2.9",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.9.0",
    "pylint==3.3.1",
    "pytest==8.3.2",
    "pytest-django==4.9.0",
    "python-crontab==3.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "redis==5.0.8",
    "rest-framework-simplejwt==0.0.2",
    "setuptools>=80.8.0",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "tomlkit==0.13.2",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
    "uvicorn==0.30.6",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "whitenoise==6.7.0",
]

[tool.uv.sources]
black = { git = "https://github.com/psf/black", rev = "a22b1ebbfd61fdc4f59de373abcc02c4db364359" }
